+++ src/main/java/com/web/flower/security/auth/JwtAuthenticationFilter.java
@@ -4,8 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.web.flower.domain.message.Message;
 import com.web.flower.domain.refresh_token.entity.RefreshToken;
 import com.web.flower.domain.refresh_token.repository.RefreshTokenRepository;
-import com.web.flower.domain.refresh_token.service.JwtService;
 import com.web.flower.domain.user.entity.User;
+import com.web.flower.utils.CookieUtils;
+import com.web.flower.utils.JwtUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -19,32 +20,27 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Optional;
 import java.util.UUID;
 
-// 스프링 시큐리티에서 UsernamePasswordAuthenticationFilter 가 있다.
-// /login 요청해서 username, password POST 전송하면 위 필터가 동작한다.
 public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
 
     private AuthenticationManager authenticationManager;
     private RefreshTokenRepository refreshTokenRepository;
-    private JwtService jwtService;
+    private JwtUtils jwtUtils;
 
-    public JwtAuthenticationFilter(AuthenticationManager authenticationManager, RefreshTokenRepository refreshTokenRepository, JwtService jwtService) {
+    public JwtAuthenticationFilter(AuthenticationManager authenticationManager, RefreshTokenRepository refreshTokenRepository, JwtUtils jwtUtils) {
         this.authenticationManager = authenticationManager;
         this.refreshTokenRepository = refreshTokenRepository;
-        this.jwtService = jwtService;
+        this.jwtUtils = jwtUtils;
         setFilterProcessesUrl("/login");
     }
 
     @Override
     public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
-        System.out.println("=== [AuthenticationFilter] 인증(로그인) 요청 도착 ===");
-
         try {
             ObjectMapper om = new ObjectMapper();
             User userEntity = om.readValue(request.getInputStream(), User.class);
@@ -62,12 +58,11 @@ public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilte
     // attemptAuthentication 실행 후, 인증이 정상적으로 되었다면 successfulAuthentication 함수가 실행된다.
     @Override
     protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authResult) throws IOException, ServletException {
-        System.out.println("=== 사용자 인증 성공 ===");
-        System.out.println("=== successfulAuthentication() ===");
+
         User userEntity = ((PrincipalDetails) authResult.getPrincipal()).getUser();
 
-        String accessToken = jwtService.createAccessToken(userEntity);
-        String refreshToken = jwtService.createRefreshToken(userEntity);
+        String accessToken = jwtUtils.createAccessToken(userEntity);
+        String refreshToken = jwtUtils.createRefreshToken(userEntity);
 
         // refresh Token DB에 저장. 이미 존재하는 RefreshToken 이 있다면, 제거 후 생성
         Optional<RefreshToken> findToken = refreshTokenRepository.findByUserId(userEntity.getId());
@@ -81,12 +76,8 @@ public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilte
                 .userId(userEntity.getId())
                 .build();
         refreshTokenRepository.save(buildToken);
-        System.out.println("refreshToken = " + refreshToken);
 
-        Cookie cookie = new Cookie("flower_token",accessToken);
-        cookie.setMaxAge(60*60*3); // 3시간
-        cookie.setHttpOnly(true);
-        response.addCookie(cookie);
+        CookieUtils.makeCookie(accessToken, response);
 
         Message message = Message.builder()
                 .status(HttpStatus.OK)
@@ -101,8 +92,6 @@ public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilte
 
     @Override
     protected void unsuccessfulAuthentication(HttpServletRequest request, HttpServletResponse response, AuthenticationException failed) throws IOException, ServletException {
-        System.out.println("=== unsuccessfulAuthentication 실행됨 =====");
-
         Object exceptionType = failed.getClass();
 
         if(exceptionType.equals(BadCredentialsException.class) || exceptionType.equals(UsernameNotFoundException.class)){
